+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Version 6.0.4-alpha (not yet released)
 
 **Improvements and fixes**
-- Main screen - Improvements to some cut off texts
+- Main screen - Improvements to some cut off texts with different display sizes
 - Main screen - Improve the “two blocks per row” threshold when using custom font scale
 - Details - Fix precipitation probability details being expressed in precipitation unit instead of %
 - Fix missing normals every other refresh